fix: add decorator to storybook preview to fix building
