--- frontend/.storybook/preview.js
@@ -1,15 +1,0 @@
-/** @type { import('@storybook/react').Preview } */
-import '../src/style/globals.css'
-
-const preview = {
-    parameters: {
-        controls: {
-            matchers: {
-                color: /(background|color)$/i,
-                date: /Date$/i,
-            },
-        },
-    },
-}
-
-export default preview
