+++ frontend/src/components/Header.stories.js
@@ -1,6 +1,5 @@
 import Header from "./Header"
 import { fn } from "@storybook/test"
-import { BrowserRouter } from "react-router-dom"
 
 export default {
   title: "Header",
@@ -8,13 +7,6 @@ export default {
   args: {
     searchHandler: fn(),
   },
-  decorators: [
-    Story => (
-      <BrowserRouter>
-        <Story />
-      </BrowserRouter>
-    ),
-  ],
 }
 
 export const DefaultHeader = {}
